+++ android/KMAPro/kMAPro/src/main/java/com/tavultesoft/kmapro/BookmarksActivity.java
@@ -23,7 +23,7 @@ import android.view.View.OnClickListener;
 import android.widget.AdapterView;
 import android.widget.BaseAdapter;
 import android.widget.EditText;
-import android.widget.ImageButton;
+import android.support.design.widget.FloatingActionButton;
 import android.widget.ListAdapter;
 import android.widget.ListView;
 import android.widget.PopupMenu;
@@ -32,14 +32,14 @@ import android.widget.TextView;
 import android.widget.Toast;
 import android.widget.AdapterView.OnItemLongClickListener;
 import android.annotation.SuppressLint;
-import android.app.ActionBar;
-import android.app.Activity;
-import android.app.AlertDialog;
+import android.support.v7.widget.Toolbar;
+import android.support.v7.app.AppCompatActivity;
+import android.support.v7.app.AlertDialog;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
 
-public class BookmarksActivity extends Activity {
+public class BookmarksActivity extends AppCompatActivity {
 
   private static ListView listView;
   private static ArrayList<HashMap<String, String>> list = null;
@@ -57,17 +57,17 @@ public class BookmarksActivity extends Activity {
     final Context context = this;
     final String title = getIntent().getStringExtra(titleKey);
     final String url = getIntent().getStringExtra(urlKey);
-    final ActionBar actionBar = getActionBar();
-    actionBar.setLogo(null);
-    actionBar.setDisplayShowHomeEnabled(false);
-    actionBar.setDisplayShowTitleEnabled(false);
-    actionBar.setDisplayShowCustomEnabled(true);
-    actionBar.setBackgroundDrawable(MainActivity.getActionBarDrawable(this));
-    final ViewGroup bookmarksTitleLayout = (ViewGroup) getLayoutInflater().inflate(
-      R.layout.bookmarks_title_layout,
-      null);
-    actionBar.setCustomView(bookmarksTitleLayout);
+
     setContentView(R.layout.bookmarks_list_layout);
+
+    final Toolbar toolbar = (Toolbar) findViewById(R.id.bookmarks_toolbar);
+    setSupportActionBar(toolbar);
+    getSupportActionBar().setLogo(null);
+    getSupportActionBar().setDisplayShowHomeEnabled(false);
+    getSupportActionBar().setDisplayShowTitleEnabled(false);
+    getSupportActionBar().setDisplayShowCustomEnabled(true);
+    getSupportActionBar().setBackgroundDrawable(MainActivity.getActionBarDrawable(this));
+
     listView = (ListView) findViewById(R.id.listView);
     list = getBookmarksList();
     String[] from = new String[]{titleKey, urlKey, iconKey};
@@ -112,7 +112,7 @@ public class BookmarksActivity extends Activity {
     TextView emptyText = (TextView) findViewById(android.R.id.empty);
     listView.setEmptyView(emptyText);
 
-    ImageButton addButton = (ImageButton) findViewById(R.id.addButton);
+    FloatingActionButton  addButton = (FloatingActionButton) findViewById(R.id.addButton);
     addButton.setOnClickListener(new OnClickListener() {
       @Override
       public void onClick(View v) {