+++ android/Samples/KMSample1/app/src/main/java/com/keyman/kmsample1/MainActivity.java
@@ -1,6 +1,6 @@
 package com.keyman.kmsample1;
 
-import android.app.Activity;
+import android.support.v7.app.AppCompatActivity;
 import android.content.res.Configuration;
 import android.os.Bundle;
 import android.util.Log;
@@ -18,12 +18,13 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-public class MainActivity extends Activity implements OnKeyboardEventListener, OnKeyboardDownloadEventListener {
+public class MainActivity extends AppCompatActivity implements OnKeyboardEventListener, OnKeyboardDownloadEventListener {
 
   private KMTextView textView;
 
   @Override
   protected void onCreate(Bundle savedInstanceState) {
+    setTheme(R.style.AppTheme);
     super.onCreate(savedInstanceState);
 
     KMManager.setDebugMode(true);