+++ android/history.md
@@ -7,6 +7,7 @@
 * Change default keyboard from `european` to `sil_euro_latin` (#1112)
 * Remove deprecated ad-hoc distribution of keyboards via `keyman://` protocol (#1109)
 * Add splash screen (#1151)
+* Update app to use Material Design theme
 
 ## 2018-08-23 10.0.505 stable
 * Validate keyboard ID when downloading keyboard from keman cloud (#1121)